refactor(registry): adjust registry structure to avoid circular reference
