+++ packages/g6/src/preset.ts
@@ -1,5 +1,5 @@
 import { runtime } from '@antv/g';
-import { registerBuiltInExtensions } from './registry';
+import { registerBuiltInExtensions } from './registry/build-in';
 
 runtime.enableCSSParsing = false;
 registerBuiltInExtensions();
