+++ packages/g6/src/registry/store.ts
@@ -1,0 +1,19 @@
+import type { ExtensionRegistry } from './types';
+
+/**
+ * <zh/> 扩展注册表
+ *
+ * <en/> Extension registry
+ */
+export const EXTENSION_REGISTRY: ExtensionRegistry = {
+  animation: {},
+  behavior: {},
+  combo: {},
+  edge: {},
+  layout: {},
+  node: {},
+  palette: {},
+  theme: {},
+  plugin: {},
+  transform: {},
+};
